# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Meta API",
    "cersion": "1.0.0",
    "description": "Rest API documentation",
    "version": "1.0.0"
  },
  "host": "hex-post-team-api-server.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "請加上 API Token"
    }
  },
  "paths": {
    "/api/posts/": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "取得依分頁取得貼文",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "分頁數",
            "type": "string"
          },
          {
            "name": "q",
            "in": "query",
            "description": "查詢",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "排序 ( arc || desc ) ",
            "type": "string"
          },
          {
            "name": "like",
            "in": "query",
            "description": "user id 查看我按讚的文章",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "取得分頁資料",
            "schema": {
              "$ref": "#/definitions/PostPage"
            }
          }
        }
      },
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "新增一般貼文 (對應設計8頁)OK",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "新增貼文",
            "schema": {
              "type": "object",
              "properties": {
                "tags": {
                  "type": "string",
                  "example": ""
                },
                "image": {
                  "type": "string",
                  "example": "https://i.imgur.com/ffCSS4H_d.webp?maxwidth=520&shape=thumb&fidelity=high"
                },
                "content": {
                  "type": "string",
                  "example": "新增貼文"
                }
              },
              "required": [
                "content"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "string",
                      "example": "628f8c9bb1e02d4b681c8fe9"
                    },
                    "tags": {
                      "type": "array",
                      "example": [
                        ""
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "type": {
                      "type": "string",
                      "example": "person"
                    },
                    "image": {
                      "type": "string",
                      "example": "https://i.imgur.com/ffCSS4H_d.webp?maxwidth=520&shape=thumb&fidelity=high"
                    },
                    "content": {
                      "type": "string",
                      "example": "https://i.imgur.com/ffCSS4H_d.webp?maxwidth=520&shape=thumb&fidelity=high"
                    },
                    "likes": {
                      "type": "number",
                      "example": 0
                    },
                    "comments": {
                      "type": "number",
                      "example": 0
                    },
                    "_id": {
                      "type": "string",
                      "example": "6290ef2bbb78cbd2b7693e5c"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-05-27T15:32:59.325Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "2022-05-27T15:32:59.325Z"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "新增成功"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "刪除全部貼文OK",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/user/": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "取得個人頁面-設定(對應設計18頁)   *",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "628f8c9bb1e02d4b681c8fe9"
                        },
                        "name": {
                          "type": "string",
                          "example": "Ray"
                        },
                        "email": {
                          "type": "string",
                          "example": "123@123.com"
                        },
                        "photo": {
                          "type": "string",
                          "example": "123"
                        },
                        "gender": {
                          "type": "string",
                          "example": "male"
                        },
                        "birthday": {
                          "type": "string",
                          "example": "2022-01-01T00:00:00.000Z"
                        },
                        "delflag": {
                          "type": "boolean",
                          "example": false
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2022-05-26T14:20:11.616Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2022-05-26T14:20:11.616Z"
                        }
                      }
                    },
                    "postCount": {
                      "type": "number",
                      "example": 0
                    },
                    "follow": {
                      "type": "number",
                      "example": 0
                    },
                    "privatepost": {
                      "type": "number",
                      "example": 0
                    },
                    "order": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "取得個人頁面資料"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "註冊使用者",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Ray"
                },
                "email": {
                  "type": "string",
                  "example": "123@123.com"
                },
                "photo": {
                  "type": "string",
                  "example": ""
                },
                "birthday": {
                  "type": "string",
                  "example": "2022-01-01"
                },
                "gender": {
                  "type": "string",
                  "example": "male || female"
                },
                "password": {
                  "type": "string",
                  "example": "12345678"
                },
                "confirmPassword": {
                  "type": "string",
                  "example": "12345678"
                }
              },
              "required": [
                "name",
                "email",
                "password",
                "confirmPassword"
              ]
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/follow/": {
      "get": {
        "tags": [
          "follow"
        ],
        "summary": "取得個人追蹤名單",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "6291e520cbbe05860f8d8c15"
                      },
                      "user": {
                        "type": "string",
                        "example": "628f8c9bb1e02d4b681c8fe9"
                      },
                      "followuser": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "628f8b08634ef4666583501b"
                            },
                            "name": {
                              "type": "string",
                              "example": "Ray"
                            },
                            "photo": {
                              "type": "string",
                              "example": "1"
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "2022-05-26T14:13:28.064Z"
                            }
                          }
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2022-05-28T09:02:24.177Z"
                      },
                      "__v": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "取得所有個人追蹤"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "follow"
        ],
        "summary": "加入追蹤(對應設計4頁)",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "schema": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "6291e520cbbe05860f8d8c15"
                    },
                    "user": {
                      "type": "string",
                      "example": "628f8c9bb1e02d4b681c8fe9"
                    },
                    "followuser": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "628f8b08634ef4666583501b"
                          },
                          "name": {
                            "type": "string",
                            "example": "Ray"
                          },
                          "photo": {
                            "type": "string",
                            "example": "1"
                          },
                          " ": {
                            "type": "string",
                            "example": "2022-05-26T14:13:28.064Z"
                          }
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-05-28T09:02:24.177Z"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "建立個人追蹤"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/email/": {
      "post": {
        "tags": [
          "email"
        ],
        "summary": "發信服務",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add a email template",
            "schema": {
              "$ref": "#/definitions/AddEmail"
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/payment/": {
      "get": {
        "tags": [
          "payment"
        ],
        "summary": "跳轉付款頁面",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "prod",
            "in": "query",
            "description": "產品ID",
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/wallet/": {
      "get": {
        "tags": [
          "wallet"
        ],
        "summary": "取得當前餘額",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Post": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "6274caa22a6bd4d1ecd8af05"
        },
        "tags": {
          "type": "array",
          "example": [
            "旅遊"
          ],
          "items": {
            "type": "string"
          }
        },
        "type": {
          "type": "string",
          "example": "group"
        },
        "image": {
          "type": "string",
          "example": "https://unsplash.com/photos/gKXKBY-C-Dk"
        },
        "content": {
          "type": "string",
          "example": "今天去看貓"
        },
        "likes": {
          "type": "number",
          "example": 99
        },
        "comments": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "type",
        "content"
      ]
    },
...
    "PostPage": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Post"
        },
        "pagination": {
          "type": "object",
          "properties": {
            "total_pages": {
              "type": "number",
              "example": 1
            },
            "current_page": {
              "type": "number",
              "example": 1
            },
            "has_pre": {
              "type": "boolean",
              "example": false
            },
            "has_next": {
              "type": "boolean",
              "example": false
            }
          }
        }
      }
    },
    "AddEmail": {
      "type": "object",
      "properties": {
        "to": {
          "type": "string",
          "example": "test@gmail.com"
        },
        "subject": {
          "type": "string",
          "example": "主旨"
        },
        "html": {
          "type": "string",
          "example": "<b>Hello world?</b>"
        },
        "text": {
          "type": "string",
          "example": "Hello world?"
        }
      }
    },
...
  }
}